# contracts/openapi/powerflow.openapi.json
{
  "components": {
    "schemas": {
...
      "BessConfig": {
        "additionalProperties": false,
        "properties": {
          "battery": {
            "$ref": "#/components/schemas/BatteryConfig"
          },
          "collector": {
            "default": "mv1",
            "title": "Collector",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Id",
            "type": "string"
          },
          "inverter": {
            "$ref": "#/components/schemas/BessInverterConfig"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "transformer": {
            "$ref": "#/components/schemas/TransformerConfig"
          }
        },
        "required": [
          "id",
          "inverter",
          "battery",
          "transformer"
        ],
        "title": "BessConfig",
        "type": "object"
      },
...
      "DefaultsRestoreResult": {
        "properties": {
          "active_site_reloaded": {
            "title": "Active Site Reloaded",
            "type": "boolean"
          },
          "maps_skipped": {
            "items": {
              "type": "string"
            },
            "title": "Maps Skipped",
            "type": "array"
          },
          "maps_written": {
            "items": {
              "type": "string"
            },
            "title": "Maps Written",
            "type": "array"
          },
          "overwrite": {
            "title": "Overwrite",
            "type": "boolean"
          },
          "sites_skipped": {
            "description": "Already stored; pass overwrite=true.",
            "items": {
              "type": "string"
            },
            "title": "Sites Skipped",
            "type": "array"
          },
          "sites_written": {
            "items": {
              "type": "string"
            },
            "title": "Sites Written",
            "type": "array"
          }
        },
        "required": [
          "overwrite",
          "sites_written",
          "sites_skipped",
          "maps_written",
          "maps_skipped",
          "active_site_reloaded"
        ],
        "title": "DefaultsRestoreResult",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "ModbusMapSummary": {
        "properties": {
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "orphaned": {
            "description": "True if the site no longer has this asset.",
            "title": "Orphaned",
            "type": "boolean"
          },
          "points": {
            "title": "Points",
            "type": "integer"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "integer"
          }
        },
        "required": [
          "asset",
          "unit_id",
          "port",
          "points",
          "orphaned"
        ],
        "title": "ModbusMapSummary",
        "type": "object"
      },
...
      "ProfileFolder": {
        "enum": [
          "load",
          "pv"
        ],
        "title": "ProfileFolder",
        "type": "string"
      },
...
      "ProfileSaveResult": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "folder": {
            "$ref": "#/components/schemas/ProfileFolder"
          },
          "reloaded_assets": {
            "description": "Active-site assets now using the new data.",
            "items": {
              "type": "string"
            },
            "title": "Reloaded Assets",
            "type": "array"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "scenario": {
            "title": "Scenario",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "folder",
          "scenario",
          "rows",
          "start",
          "end",
          "columns",
          "reloaded_assets"
        ],
        "title": "ProfileSaveResult",
        "type": "object"
      },
...
      "SiteConfig": {
        "additionalProperties": false,
        "description": "A grid-connected site: the utility feeder, the POI, MV collectors and N BESS/PV/loads.",
        "properties": {
          "bess": {
            "items": {
              "$ref": "#/components/schemas/BessConfig"
            },
            "title": "Bess",
            "type": "array"
          },
          "collectors": {
            "items": {
              "$ref": "#/components/schemas/CollectorConfig"
            },
            "minItems": 1,
            "title": "Collectors",
            "type": "array"
          },
          "grid": {
            "$ref": "#/components/schemas/GridConfig"
          },
          "interfaces": {
            "$ref": "#/components/schemas/InterfacesConfig"
          },
          "loads": {
            "items": {
              "$ref": "#/components/schemas/LoadConfig"
            },
            "title": "Loads",
            "type": "array"
          },
          "poi": {
            "$ref": "#/components/schemas/PoiConfig"
          },
          "pv": {
            "items": {
              "$ref": "#/components/schemas/PvConfig"
            },
            "title": "Pv",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "simulation": {
            "$ref": "#/components/schemas/SimulationConfig"
          },
          "site": {
            "$ref": "#/components/schemas/SiteInfo"
          }
        },
        "title": "SiteConfig",
        "type": "object"
      },
      "SiteInfo": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "default": "Site",
            "title": "Name",
            "type": "string"
          }
        },
        "title": "SiteInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Grid-connected microgrid power flow simulator (pandapower, balanced positive sequence,\nquasi-static). The EMS writes setpoints; every step the simulator applies the asset limits,\nsolves the network and publishes a measurement snapshot.\n\n**Sign conventions**\n- BESS / PV P and Q: generator convention. P > 0 injects (BESS discharging, PV producing),\n  P < 0 is a charging BESS. Q > 0 injects vars.\n- Load P and Q: P > 0 consumes, Q > 0 consumes vars (lagging).\n- POI: P > 0 exports to the utility, P < 0 imports. The same for Q.\n- Transformers: P/Q > 0 flows toward the grid (LV → HV).\n- pf = |P|/S, signed with Q in the element's own convention.\n\n**Units:** kW, kvar, kVA, kWh, kV (line-to-line), pu where noted, s.\n\n**Points:** `<asset_type>.<asset_id>.<point>`, listed at GET /api/points.\n",
    "title": "powerflow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/defaults/restore": {
      "post": {
        "description": "Without overwrite, sites and maps already stored are kept (reported as skipped). With overwrite=true they're replaced; the simulation must be stopped (409), and the active site is reloaded if it was replaced. The active-site choice is unchanged.",
        "operationId": "restore_defaults_api_defaults_restore_post",
        "parameters": [
          {
            "in": "query",
            "name": "overwrite",
            "required": false,
            "schema": {
              "default": false,
              "title": "Overwrite",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultsRestoreResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Import the default sites and maps into the database",
        "tags": [
          "defaults"
        ]
      }
    },
...
    "/api/profiles/{folder}/{scenario}": {
      "delete": {
        "operationId": "delete_profile_api_profiles__folder___scenario__delete",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProfileFolder"
            }
          },
          {
            "in": "path",
            "name": "scenario",
            "required": true,
            "schema": {
              "title": "Scenario",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Delete a profile scenario (409 while any stored site uses it)",
        "tags": [
          "profiles"
        ]
      },
      "get": {
        "operationId": "get_profile_api_profiles__folder___scenario__get",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProfileFolder"
            }
          },
          {
            "in": "path",
            "name": "scenario",
            "required": true,
            "schema": {
              "title": "Scenario",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "A profile scenario's CSV",
        "tags": [
          "profiles"
        ]
      },
      "put": {
        "description": "Validated before it's written. Active-site assets using the scenario pick it up on the next step, even while running (their configured scale/loop still apply).",
        "operationId": "put_profile_api_profiles__folder___scenario__put",
        "parameters": [
          {
            "in": "path",
            "name": "folder",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProfileFolder"
            }
          },
          {
            "in": "path",
            "name": "scenario",
            "required": true,
            "schema": {
              "title": "Scenario",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          },
          "description": "Load: timestamp,p_kw,q_kvar (or timestamp,p_kw,pf). PV: timestamp and p_kw (ac_kw source) and/or ghi_wm2 (irradiance source). Timestamps ISO 8601.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSaveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Create or replace a profile scenario (CSV body)",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/api/schemas/{name}": {
      "get": {
        "operationId": "get_schema_api_schemas__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/JsonValue"
                  },
                  "title": "Response Get Schema Api Schemas  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "One JSON Schema, e.g. site-config or modbus-map",
        "tags": [
          "schemas"
        ]
      }
    },
...
    "/api/sites/{name}": {
      "delete": {
        "operationId": "delete_site_api_sites__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Delete a site",
        "tags": [
          "sites"
        ]
      },
      "get": {
        "operationId": "get_site_api_sites__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Get Site",
        "tags": [
          "sites"
        ]
      },
      "put": {
        "description": "Validated (schema, references, profile scenarios exist) before it's written. Saving the active site reloads it, so the simulation must be stopped (409 otherwise).",
        "operationId": "put_site_api_sites__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Create or replace a stored site",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{name}/activate": {
      "post": {
        "description": "Rebuilds the network, resets the simulation and records the site as the one to load at startup. The engine stays stopped: POST /api/sim/start.",
        "operationId": "activate_site_api_sites__name__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "Load a stored site (only while stopped)",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site}/modbus-maps": {
      "get": {
        "operationId": "list_maps_api_sites__site__modbus_maps_get",
        "parameters": [
          {
            "in": "path",
            "name": "site",
            "required": true,
            "schema": {
              "title": "Site",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModbusMapSummary"
                  },
                  "title": "Response List Maps Api Sites  Site  Modbus Maps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such site, map or scenario."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed now (running, or in use)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid name or content."
          }
        },
        "summary": "A site's per-asset Modbus maps",
        "tags": [
          "modbus maps"
        ]
      }
    },
...
  }
}